backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.38.32",
    "celery[redis]>=5.5.3",
    "django>=5.2.2",
    "djangorestframework>=3.16.0",
    "drf-spectacular>=0.28.0",
    "oci>=2.154.0",
    "python-dotenv>=1.1.0",
]
